Remove unused legacy controllers & services.
