--- app/Http/Requests/Legacy/Two/TagsRequest.php
@@ -1,31 +1,0 @@
-<?php
-
-namespace Rogue\Http\Requests\Legacy\Two;
-
-use Rogue\Http\Requests\Request;
-
-class TagsRequest extends Request
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            'post_id' => 'required|int',
-            'tag_name' => 'required|string',
-        ];
-    }
-}
