+++ tests/Unit/Services/CampaignServiceTest.php
@@ -5,10 +5,7 @@ namespace Tests\Unit\Services;
 use Tests\TestCase;
 use Rogue\Models\Post;
 use Rogue\Models\Signup;
-use Rogue\Services\Ashes;
-use Rogue\Services\Phoenix;
 use Rogue\Services\CampaignService;
-use Rogue\Repositories\CacheRepository;
 
 class CampaignServiceTest extends TestCase
 {
